document_ai/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async task queue
celery[redis]==5.3.6
redis==5.0.1

# Document parsing
unstructured==0.12.5
pdf2image==1.17.0
pypdf==4.0.1
python-docx==1.1.0

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# LLM and AI
openai==1.12.0
transformers==4.37.0
torch==2.2.0

# NLP and entity extraction
spacy==3.8.0
nltk==3.8.1

# Sentiment analysis
textblob==0.17.1
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# Google Cloud Vision (optional)
google-cloud-vision==3.4.5

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Storage
boto3==1.34.0
